Split off the conversion from blender/python objects into strings to another file. More pylint
